Move API handlers to match URL hierarchy
