+++ requirements.txt
@@ -11,11 +11,11 @@ aplpy==1.1.1
 avro-python3==1.8.2
 fastavro==0.21.7
 tqdm==4.23.2
-matplotlib==2.1.0
+matplotlib>=3
 astroquery==0.3.8
 sqlalchemy-utils
 apispec==2.0.2
 marshmallow==3.0.0rc5
 marshmallow-sqlalchemy==0.17.0
 marshmallow-enum==1.4.1
-Pillow==6.2.0
+Pillow>=6
