+++ skyportal/handlers/__init__.py
@@ -3,20 +3,7 @@ from baselayer.app.handlers import (MainPageHandler,
                                     LogoutHandler)
 from baselayer.app.custom_exceptions import AccessError
 
-from .source import SourceHandler, FilterSourcesHandler, SourcePhotometryHandler
-from .comment import CommentHandler
-from .group import GroupHandler, GroupUserHandler
-from .plot import PlotPhotometryHandler, PlotSpectroscopyHandler
-from .profile import ProfileHandler
-from .logout import LogoutHandler
-from .become_user import BecomeUserHandler
-from .photometry import PhotometryHandler
-from .token import TokenHandler
-from .sysinfo import SysInfoHandler
-from .user import UserHandler
-from .spectrum import SpectrumHandler
-from .thumbnail import ThumbnailHandler
-from .dbinfo import DBInfoHandler
 from .base import BaseHandler
-from .instrument import InstrumentHandler
-from .telescope import TelescopeHandler
+
+from .become_user import BecomeUserHandler
+from .logout import LogoutHandler
