+++ skyportal/handlers/api/__init__.py
@@ -1,0 +1,10 @@
+from .comment import CommentHandler
+from .group import GroupHandler, GroupUserHandler
+from .instrument import InstrumentHandler
+from .photometry import PhotometryHandler
+from .source import SourceHandler, FilterSourcesHandler, SourcePhotometryHandler
+from .spectrum import SpectrumHandler
+from .sysinfo import SysInfoHandler
+from .telescope import TelescopeHandler
+from .thumbnail import ThumbnailHandler
+from .user import UserHandler
