+++ skyportal/handlers/api/internal/__init__.py
@@ -1,0 +1,4 @@
+from .plot import PlotPhotometryHandler, PlotSpectroscopyHandler
+from .token import TokenHandler
+from .dbinfo import DBInfoHandler
+from .profile import ProfileHandler
