+++ skyportal/handlers/api/internal/token.py
@@ -1,7 +1,7 @@
-from .base import BaseHandler
+from ...base import BaseHandler
 from baselayer.app.access import permissions, auth_or_token
-from ..models import User, Token, DBSession
-from ..model_util import create_token
+from ....models import User, Token, DBSession
+from ....model_util import create_token
 
 import tornado.web
