+++ skyportal/handlers/api/spectrum.py
@@ -1,19 +1,21 @@
+import tornado.web
+from sqlalchemy.orm import joinedload
 from marshmallow.exceptions import ValidationError
 from baselayer.app.access import permissions, auth_or_token
-from .base import BaseHandler
-from ..models import DBSession, Instrument, Telescope
+from ..base import BaseHandler
+from ...models import DBSession, Spectrum, Comment, Instrument, Source
 
 
-class InstrumentHandler(BaseHandler):
+class SpectrumHandler(BaseHandler):
     @permissions(['Upload data'])
     def post(self):
         """
         ---
-        description: Create instruments
+        description: Upload spectrum
         parameters:
           - in: path
-            name: instrument
-            schema: Instrument
+            name: spectrum
+            schema: Spectrum
         responses:
           200:
             content:
@@ -25,34 +27,35 @@ class InstrumentHandler(BaseHandler):
                       properties:
                         id:
                           type: integer
-                          description: New instrument ID
+                          description: New spectrum ID
         """
         data = self.get_json()
-        telescope_id = data.pop('telescope_id')
-        telescope = Telescope.get_if_owned_by(telescope_id, self.current_user)
-        if not telescope:
-            return self.error('Invalid telescope ID.')
+        source_id = data.pop('source_id')
+        instrument_id = data.pop('instrument_id')
+        source = Source.get_if_owned_by(source_id, self.current_user)
+        instrument = Instrument.query.get(instrument_id)
 
-        schema = Instrument.__schema__()
+        schema = Spectrum.__schema__()
         try:
-            instrument = schema.load(data)
+            spec = schema.load(data)
         except ValidationError as e:
             return self.error('Invalid/missing parameters: '
                               f'{e.normalized_messages()}')
-        instrument.telescope = telescope
-        DBSession.add(instrument)
+        spec.source = source
+        spec.instrument = instrument
+        DBSession().add(spec)
         DBSession().commit()
 
-        return self.success(data={"id": instrument.id})
+        return self.success(data={"id": spec.id})
 
     @auth_or_token
-    def get(self, instrument_id):
+    def get(self, spectrum_id):
         """
         ---
-        description: Retrieve an instrument
+        description: Retrieve a spectrum
         parameters:
           - in: path
-            name: instrument_id
+            name: spectrum_id
             required: true
             schema:
               type: integer
@@ -60,31 +63,30 @@ class InstrumentHandler(BaseHandler):
           200:
             content:
               application/json:
-                schema: SingleInstrument
+                schema: SingleSpectrum
           400:
             content:
               application/json:
                 schema: Error
         """
-        instrument = Instrument.query.get(int(instrument_id))
+        spectrum = Spectrum.query.get(spectrum_id)
 
-        if instrument is not None:
-            telescope = Telescope.get_if_owned_by(instrument.telescope_id,
-                                                  self.current_user)
-            return self.success(data={'instrument': instrument})
+        if spectrum is not None:
+            source = Source.get_if_owned_by(spectrum.source_id, self.current_user)
+            return self.success(data={'spectrum': spectrum})
         else:
-            return self.error(f"Could not load instrument {instrument_id}",
-                              data={"instrument_id": instrument_id})
+            return self.error(f"Could not load spectrum {spectrum_id}",
+                              data={"spectrum_id": spectrum_id})
 
     @permissions(['Manage sources'])
-    def put(self, instrument_id):
+    def put(self, spectrum_id):
         """
         ---
-        description: Update instrument
+        description: Update spectrum
         parameters:
           - in: path
-            name: instrument
-            schema: Instrument
+            name: spectrum
+            schema: Spectrum
         responses:
           200:
             content:
@@ -95,13 +97,12 @@ class InstrumentHandler(BaseHandler):
               application/json:
                 schema: Error
         """
-        instrument = Instrument.query.get(int(instrument_id))
-        telescope = Telescope.get_if_owned_by(instrument.telescope_id,
-                                              self.current_user)
+        spectrum = Spectrum.query.get(spectrum_id)
+        source = Source.get_if_owned_by(spectrum.source_id, self.current_user)
         data = self.get_json()
-        data['id'] = int(instrument_id)
+        data['id'] = spectrum_id
 
-        schema = Instrument.__schema__()
+        schema = Spectrum.__schema__()
         try:
             schema.load(data)
         except ValidationError as e:
@@ -112,13 +113,13 @@ class InstrumentHandler(BaseHandler):
         return self.success()
 
     @permissions(['Manage sources'])
-    def delete(self, instrument_id):
+    def delete(self, spectrum_id):
         """
         ---
-        description: Delete an instrument
+        description: Delete a spectrum
         parameters:
           - in: path
-            name: instrument_id
+            name: spectrum_id
             required: true
             schema:
               type: integer
@@ -132,10 +133,9 @@ class InstrumentHandler(BaseHandler):
               application/json:
                 schema: Error
         """
-        instrument = Instrument.query.get(int(instrument_id))
-        telescope = Telescope.get_if_owned_by(instrument.telescope_id,
-                                              self.current_user)
-        DBSession.query(Instrument).filter(Instrument.id == int(instrument_id)).delete()
+        spectrum = Spectrum.query.get(spectrum_id)
+        source = Source.get_if_owned_by(spectrum.source_id, self.current_user)
+        DBSession().delete(spectrum)
         DBSession().commit()
 
         return self.success()
